Tied SC2 client events to Overwolf app connector.

+++ apps/toolkit-app/src/main/java/sc2toolkit/app/toolkit/overwolf/OverwolfAppConnector.java
@@ -22,7 +22,7 @@ public interface OverwolfAppConnector {
    * @param displayTime The display time.
    * @return A {@link CompletionStage} for the operation result.
    */
-  CompletionStage<Void> updateGameTime(float displayTime);
+  CompletionStage<Void> updateGameTime(double displayTime);
 
   /**
    * Sends game end information.
